put filtering logic in api call
